requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1
Werkzeug==2.3.7
PyMySQL==1.1.0
python-dotenv==1.0.0
bcrypt==4.0.1
email-validator==2.0.0
reportlab==4.0.7
Flask-Mail==0.9.1
gunicorn==21.2.0
whitenoise==6.6.0
